+++ Mechanics/Common.lua
@@ -33,6 +33,17 @@ function MyDungeonsBook:FindNameDeclension(petTooltipText, partyMemberName)
 	return false;
 end
 
+--[[
+@local
+@return[type=bool]
+]]
+function MyDungeonsBook:IsFriendlyFire(spellId)
+	if (spellId == 98021) then
+		return true;
+	end
+	return false;
+end
+
 --[[--
 Original idea is taken from Details addon
 
@@ -507,6 +518,9 @@ end
 @param[type=number] amount amount of damage done to `unit` by `spellId`
 ]]
 function MyDungeonsBook:SaveTrackedDamageToPartyMembers(key, unit, spellId, amount)
+	if (self:IsFriendlyFire(spellId)) then
+		return;
+	end
 	local amountInPercents = amount and amount / UnitHealthMax(unit) * 100 or 0;
 	if (amountInPercents >= 40 and self.db.global.meta.mechanics[key].verbose) then
 		local spellLink = GetSpellLink(spellId);